requirements.txt
django==1.11
python-dotenv